+++ src/Pages/ResultPage/ResultPage.js
@@ -4,19 +4,10 @@ import ResultContents from "./Components/ResultContents";
 import { useState } from "react";
 
 function ResultPage() {
-  const [result, setResult] = useState({
-    symptom1: 0,
-    symptom2: 0,
-    symptom3: 0,
-    symptom4: 0,
-    total: 0,
-    date: "",
-  });
-
   return (
     <ResultPageContainer>
       <ResultHeader />
-      <ResultContents result={result} />
+      <ResultContents />
     </ResultPageContainer>
   );
 }
@@ -27,7 +18,6 @@ const ResultPageContainer = styled.div`
   height: 100vh;
   width: 100vw;
   background-color: black;
-
   display: flex;
   justify-content: center;
 `;
